tests: Skip 'guix-package-net.sh' when prerequisites aren't met.
* tests/guix-package-net.sh: Reinstate networking and shebang test that
  inadvertently vanished in commit 57b8623.

+++ tests/guix-package-net.sh
@@ -37,6 +37,14 @@ shebang_too_long ()
 	 -ge 128
 }
 
+if ! guile -c '(getaddrinfo "www.gnu.org" "80" AI_NUMERICSERV)' 2> /dev/null \
+	|| shebang_too_long
+then
+    # Skipping.
+    exit 77
+fi
+
+
 profile="t-profile-$$"
 rm -f "$profile"
